+++ CHANGELOG.md
@@ -120,6 +120,8 @@ We made many tests but risk zero do not exist, so if you see any bug, please rep
   - Org licenses
 - **hardis:org:diagnose:audittrail**: Add new filtered actions
   - Customer Portal: createdcustomersuccessuser
+- AI Deployment assistant
+  - Add error `Change Matching Rule`
 
 ### Fixes